Staff login

+++ app/Http/Controllers/API/ClinicController.php
@@ -1,62 +1,81 @@
 <?php
 
-namespace App\Http\Controllers\API;
+namespace App\Http\Controllers\Api;
 
-use App\Http\Controllers\Controller;
-use App\Http\Requests\Clinic\CreateClinicRequest;
-use App\Http\Requests\Clinic\UpdateClinicRequest;
-use App\Http\Resources\Clinic\ClinicResource;
 use App\Models\Clinic;
-use Illuminate\Http\JsonResponse;
-use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
+use App\Http\Controllers\Controller;
+use App\Http\Resources\ClinicResource;
+use App\Http\Resources\ClinicCollection;
+use App\Http\Requests\ClinicStoreRequest;
+use App\Http\Requests\ClinicUpdateRequest;
+use Dotenv\Exception\ValidationException;
+use Error;
 use Illuminate\Support\Facades\Auth;
-use Essa\APIToolKit\Api\ApiResponse;
 
 class ClinicController extends Controller
 {
-    use APIResponse;
-    public function __construct()
+    public function index(Request $request): ClinicCollection
     {
+        $this->authorize('view-any', Clinic::class);
 
+        $search = $request->get('search', '');
+
+        $clinics = Clinic::search($search)
+            ->latest()
+            ->paginate();
+
+        return new ClinicCollection($clinics);
     }
 
-    public function index(): AnonymousResourceCollection 
+    public function store(ClinicStoreRequest $request): ClinicResource
     {
-        $clinics = Clinic::useFilters()->dynamicPaginate();
+        // $this->authorize('create', Clinic::class);
+        $user = Auth::user();
 
-        return ClinicResource::collection($clinics);
-    }
-
-    public function store(CreateClinicRequest $request): JsonResponse
-    {
-        if (auth()->check()) {
-            $data = $request->validated();
-            $data['user_id'] = auth()->id();
-            $clinic = Clinic::create($data);
-            return $this->responseCreated('Clinic created successfully', new ClinicResource($clinic));
-        } else {
-            return response()->json(['error' => 'User not authenticated'], 401);
+        if ($user->role != 'admin') {
+            throw Error::validation('Only Admins can create Clinics', ['Only Admins can create Clinics']);
         }
+
+
+        $validated = $request->validated();
+
+        $clinic = Clinic::create($validated);
+
+        
+        // update user clinic_id to newly created clinic
+        $user->clinic_id = $request->clinic_id;
+
+        return new ClinicResource($clinic);
     }
 
-    public function show(Clinic $clinic): JsonResponse
+    public function show(Request $request, Clinic $clinic): ClinicResource
     {
-        return $this->responseSuccess(null, new ClinicResource($clinic));
+        $this->authorize('view', $clinic);
+
+        return new ClinicResource($clinic);
     }
 
-    public function update(UpdateClinicRequest $request, Clinic $clinic): JsonResponse
-    {
-        $clinic->update($request->validated());
+    public function update(
+        ClinicUpdateRequest $request,
+        Clinic $clinic
+    ): ClinicResource {
+        $this->authorize('update', $clinic);
 
-        return $this->responseSuccess('Clinic updated Successfully', new ClinicResource($clinic));
+        $validated = $request->validated();
+
+        $clinic->update($validated);
+
+        return new ClinicResource($clinic);
     }
 
-    public function destroy(Clinic $clinic): JsonResponse
+    public function destroy(Request $request, Clinic $clinic): Response
     {
+        $this->authorize('delete', $clinic);
+
         $clinic->delete();
 
-        return $this->responseDeleted();
+        return response()->noContent();
     }
-
 }
